assignment1 problems solved

+++ src/model/fragment.js
@@ -3,7 +3,7 @@
 const { randomUUID } = require('crypto');
 // Use https://www.npmjs.com/package/content-type to create/parse Content-Type headers
 const contentType = require('content-type');
-
+const logger = require('../logger');
 // Functions for working with fragment metadata/data using our DB
 const {
   readFragment,
@@ -14,28 +14,39 @@ const {
   deleteFragment,
 } = require('./data');
 
+const validTypes = {
+  txt: 'text/plain',
+  txtCharset: 'text/plain; charset=utf-8',
+  md: 'text/markdown',
+  html: 'text/html',
+  json: 'application/json',
+};
+
 class Fragment {
   constructor({ id, ownerId, created, updated, type, size = 0 }) {
     // TODO
-    if (!ownerId || !type) {
-      throw new Error('Owner ID and type needed to create a fragment');
-    }
     // check type and throw if unsupported type passed
-    if (!Fragment.isSupportedType(type)) {
-      throw new Error(`${type} is not supported`);
-    }
-    // check size
-    if (typeof size != 'number' || size < 0) {
-      throw new Error('Size must be a positive number');
-    }
+    if (ownerId && type) {
+      if (Fragment.isSupportedType(type)) {
+        this.type = type;
+      } else throw Error('Invalid type');
+      // check size
+      if (size) {
+        if (Number.isInteger(size) && (Math.sign(size) === 0 || Math.sign(size) === 1)) {
+          this.size = size;
+        } else throw Error('Invalid size value');
+      } else this.size = 0;
 
-    // generate id for new fragment
-    this.id = id || randomUUID();
-    this.ownerId = ownerId;
-    this.type = type;
-    this.created = created || new Date().toISOString();
-    this.updated = updated || new Date().toISOString();
-    this.size = size || 0;
+      // generate id for new fragment
+      this.id = id || randomUUID();
+      this.ownerId = ownerId;
+      if (!created || Object.keys(created).length === 0) {
+        this.created = new Date().toISOString();
+      } else this.created = created;
+      if (!updated || Object.keys(updated).length === 0) {
+        this.updated = new Date().toISOString();
+      } else this.updated = updated;
+    } else throw Error('Missing ownerId or type');
   }
 
   /**
@@ -46,6 +57,7 @@ class Fragment {
    */
   static async byUser(ownerId, expand = false) {
     // TODO
+    logger.debug({ ownerId, expand }, 'byUser()');
     return listFragments(ownerId, expand);
   }
 
@@ -57,11 +69,18 @@ class Fragment {
    */
   static async byId(ownerId, id) {
     // TODO
-    const fragment = await readFragment(ownerId, id);
-    if (!fragment) {
-      throw new Error(`Can not find fragment with id: ${id}`);
+    try {
+      const data = await readFragment(ownerId, id);
+      if (data === undefined) {
+        throw new Error('Fragment does not exist.');
+      }
+
+      logger.debug({ data }, 'byId()');
+
+      return data;
+    } catch (err) {
+      throw new Error(err);
     }
-    return new Fragment(fragment);
   }
 
   /**
@@ -72,7 +91,13 @@ class Fragment {
    */
   static delete(ownerId, id) {
     // TODO
-    return deleteFragment(ownerId, id);
+    if (ownerId && id) {
+      try {
+        return deleteFragment(ownerId, id);
+      } catch (err) {
+        throw new Error(err);
+      }
+    } else throw Error('Invalid data passed');
   }
 
   /**
@@ -91,7 +116,8 @@ class Fragment {
    */
   async getData() {
     // TODO
-    return readFragmentData(this.ownerId, this.id);
+    const data = await readFragmentData(this.ownerId, this.id);
+    return data;
   }
 
   /**
@@ -101,10 +127,22 @@ class Fragment {
    */
   async setData(data) {
     // TODO
-    this.updated = new Date().toISOString();
-    this.size = Buffer.byteLength(data);
-    this.save();
-    return writeFragmentData(this.ownerId, this.id, data);
+    try {
+      if (!data) {
+        throw Error('Data is empty');
+      }
+
+      if (!Buffer.isBuffer(data)) {
+        throw Error('Data is not buffer');
+      }
+
+      this.size = Buffer.byteLength(data);
+      this.updated = new Date().toISOString();
+
+      return await writeFragmentData(this.ownerId, this.id, data);
+    } catch (err) {
+      throw new Error(err);
+    }
   }
 
   /**
@@ -123,18 +161,37 @@ class Fragment {
    */
   get isText() {
     // TODO
-    const regex = new RegExp('^text/*');
-    return regex.test(this.type);
+    const type = this.mimeType;
+    logger.debug({ type }, 'isText()');
+    return type.includes('text');
   }
-
   /**
    * Returns the formats into which this fragment type can be converted
    * @returns {Array<string>} list of supported mime types
    */
   get formats() {
     // TODO
-    const formats = ['text/plain'];
-    return formats;
+    let array = [];
+
+    switch (this.type) {
+      case validTypes.txt:
+      case validTypes.txtCharset:
+        array = [validTypes.txt];
+        break;
+      case validTypes.md:
+        array = [validTypes.md, validTypes.txt, validTypes.html];
+        break;
+      case validTypes.html:
+        array = [validTypes.html, validTypes.txt];
+        break;
+      case validTypes.json:
+        array = [validTypes.json, validTypes.txt];
+        break;
+      default:
+        array = [];
+    }
+
+    return array;
   }
 
   /**
@@ -144,9 +201,9 @@ class Fragment {
    */
   static isSupportedType(value) {
     // TODO
-    const regex = new RegExp('^(text/*|application/json/*|image/*)');
-    return regex.test(value);
+    return Object.values(validTypes).includes(value);
   }
 }
 
 module.exports.Fragment = Fragment;
+module.exports.validTypes = validTypes;
